i965,iris/blorp: blitting of integers instead of floats, part 2
Blorp blitting uses floats for better precision.
But still its not enough for some tests.
This patch avoids usage of floats in API.
Instead integers are used.
Patch covers clipping for integers.
Part 2 is done separately because it almost independent,
but changes (incorrect type) may break functionality.
With patch-changes all current tests has to be passed
(especially fbo-blit-scaled-linear).

v2: Iris support added

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=110282
CC: Jason Ekstrand <jason.ekstrand@intel.com>
CC: Lionel G Landwerlin <lionel.g.landwerlin@intel.com>
CC: Kenneth Graunke <kenneth@whitecape.org>
Reported-by: Vadym Shovkoplias <vadym.shovkoplias@globallogic.com>

+++ src/mesa/drivers/dri/i965/brw_meta_util.h
@@ -36,10 +36,10 @@ bool
 brw_meta_mirror_clip_and_scissor(const struct gl_context *ctx,
                                  const struct gl_framebuffer *read_fb,
                                  const struct gl_framebuffer *draw_fb,
-                                 GLfloat *srcX0, GLfloat *srcY0,
-                                 GLfloat *srcX1, GLfloat *srcY1,
-                                 GLfloat *dstX0, GLfloat *dstY0,
-                                 GLfloat *dstX1, GLfloat *dstY1,
+                                 GLint *srcX0, GLint *srcY0,
+                                 GLint *srcX1, GLint *srcY1,
+                                 GLint *dstX0, GLint *dstY0,
+                                 GLint *dstX1, GLint *dstY1,
                                  GLdouble *scale_x, GLdouble *scale_y,
                                  bool *mirror_x, bool *mirror_y);
